Reorganized components/bookmarks folder into subfolders
