+++ src/shared/components/bookmarks/actionMenus/index.js
@@ -1,0 +1,10 @@
+export {
+  FileActionMenuContextProvider,
+  FileActionMenuContext
+} from "./FileActionMenuContext";
+export {
+  FolderActionMenuContextProvider,
+  FolderActionMenuContext
+} from "./FolderActionMenuContext";
+export { default as FileActionMenu } from "./FileActionMenu";
+export { default as FolderActionMenu } from "./FolderActionMenu";
