+++ src/shared/components/bookmarks/index.js
@@ -1,13 +1,6 @@
 export * from "./form";
-export { FileActionMenuContextProvider } from "./FileActionMenuContext";
-export { FolderActionMenuContextProvider } from "./FolderActionMenuContext";
+export * from "./list";
+export * from "./tree";
+export * from "./actionMenus";
 export { default as BookmarkBreadcrumbs } from "./BookmarkBreadcrumbs";
 export { default as BookmarkIcon } from "./BookmarkIcon";
-export { default as BookmarkList } from "./BookmarkList";
-export { default as BookmarkTreeView } from "./BookmarkTreeView";
-export { default as FileActionMenu } from "./FileActionMenu";
-export { default as FileListItem } from "./FileListItem";
-export { default as FileTreeItem } from "./FileTreeItem";
-export { default as FolderActionMenu } from "./FolderActionMenu";
-export { default as FolderListItem } from "./FolderListItem";
-export { default as FolderTreeItem } from "./FolderTreeItem";
