+++ src/shared/components/bookmarks/list/FileListItem.js
@@ -1,35 +1,28 @@
 import React from "react";
-import { connect } from "react-redux";
 import PropTypes from "prop-types";
 import {
   ListItem,
   ListItemIcon,
   ListItemSecondaryAction,
-  Icon,
   IconButton,
   Typography
 } from "@material-ui/core";
 import MoreVertIcon from "@material-ui/icons/MoreVert";
-import { applyFilter } from "shared/store/actions";
-import { FolderActionMenuContext } from "./FolderActionMenuContext";
+import { FileActionMenuContext } from "../actionMenus";
+import BookmarkIcon from "../BookmarkIcon";
 
-export function FolderListItem(props) {
-  const { setAnchorEl } = React.useContext(FolderActionMenuContext);
-  const { node = {}, applyFilter, iconSize = 24, selected, ...others } = props;
+export default function FileListItem(props) {
+  const { setAnchorEl } = React.useContext(FileActionMenuContext);
+  const { node = {}, iconSize = 24, selected, ...others } = props;
 
   function showActionMenu(event) {
     setAnchorEl(event.currentTarget);
   }
 
   return (
-    <ListItem
-      button
-      onDoubleClick={() => applyFilter({ parentId: node.id })}
-      style={{ minHeight: "35px" }}
-      {...others}
-    >
-      <ListItemIcon style={{ minWidth: iconSize, padding: 1 }}>
-        <Icon style={{ fontSize: iconSize + 3 }}>folder</Icon>
+    <ListItem button style={{ minHeight: "35px" }} {...others}>
+      <ListItemIcon style={{ minWidth: iconSize, margin: 2 }}>
+        <BookmarkIcon url={node.url} size={iconSize} />
       </ListItemIcon>
       <Typography
         variant="body2"
@@ -40,11 +33,23 @@ export function FolderListItem(props) {
       >
         {node.title}
       </Typography>
+      <Typography
+        variant="body2"
+        noWrap
+        component="span"
+        color="textSecondary"
+        style={{
+          marginInlineStart: "1em",
+          ...(!selected && { display: "none" })
+        }}
+      >
+        {node.url}
+      </Typography>
       <ListItemSecondaryAction>
         <IconButton
           edge="end"
           aria-label="more actions"
-          aria-controls={`folder-action-menu`}
+          aria-controls={`bookmark-action-menu`}
           aria-haspopup="true"
           onClick={showActionMenu}
           color="inherit"
@@ -57,13 +62,9 @@ export function FolderListItem(props) {
   );
 }
 
-export default connect(
-  null,
-  { applyFilter }
-)(FolderListItem);
-
-FolderListItem.propTypes = {
+FileListItem.propTypes = {
   node: PropTypes.shape({
-    title: PropTypes.string
+    title: PropTypes.string,
+    url: PropTypes.string
   }).isRequired
 };
