+++ src/shared/components/bookmarks/list/index.js
@@ -1,0 +1,3 @@
+export { default as BookmarkList } from "./BookmarkList";
+export { default as FileListItem } from "./FileListItem";
+export { default as FolderListItem } from "./FolderListItem";
