+++ src/shared/components/bookmarks/tree/index.js
@@ -1,0 +1,3 @@
+export { default as BookmarkTreeView } from "./BookmarkTreeView";
+export { default as FileTreeItem } from "./FileTreeItem";
+export { default as FolderTreeItem } from "./FolderTreeItem";
