+++ src/shared/components/helpers/index.js
@@ -1,7 +1,5 @@
 export { default as LinkTab } from "./LinkTab";
 export { default as SearchInput } from "./SearchInput";
 export { default as SnackbarCloseButton } from "./SnackbarCloseButton";
-export { default as StyledTreeItem } from "./StyledTreeItem";
 export { default as SubmitButton } from "./SubmitButton";
 export { default as TabPanel } from "./TabPanel";
-export { default as TreeItemStyles } from "./TreeItemStyles";
